# swagger.yaml
swagger: '2.0'
info:
  version: 1.0.0
  title: Node Swagger Exampol
  license:
    name: Proprietary
basePath: /v1
consumes:
  - application/json
produces:
  - application/json
schemes:
  - http
  - https
paths:
...
  '/register':
    post:
      tags:
        - Register
      operationId: Register
      parameters:
        - name: user
          in: body
          schema:
            type: object
            properties:
              name:
                type: string
              address:
                type: string
              password:
                type: string    
      description: |
        Register User
      responses:
        '200':
          description: Successful
          schema:
            type: string
    get:
      tags:
        - Register
      operationId: GetRegistee   
      description: |
        Register User      
      responses:
        '200':
          description: Successful
          schema:
            type: string
...